Crate conjure_cp_essence_parser

Re-exports

pub use errors::EssenceParseError;

pub use parser_legacy::parse_essence_file;

pub use parser::*;

Modules

diagnostics

errors

parser

parser_legacy

Macros

child

Get the i-th child of a node, or return a syntax error with a message if it doesn’t exist.

field

Get the named field of a node, or return a syntax error with a message if it doesn’t exist.

named_child

Get the i-th named child of a node, or return a syntax error with a message if it doesn’t exist.